move and rename package
