+++ pkg/services/ngalert/ngalert.go
@@ -20,6 +20,7 @@ import (
 	"github.com/grafana/grafana/pkg/services/ngalert/notifier/inhibition_rules"
 	"github.com/grafana/grafana/pkg/services/ngalert/notifier/routes"
 	"github.com/grafana/grafana/pkg/services/ngalert/provisioning/validation"
+	"github.com/grafana/grafana/pkg/services/ngalert/store/folderlabelsyncer"
 
 	"github.com/grafana/grafana/pkg/api/routing"
 	"github.com/grafana/grafana/pkg/bus"
@@ -42,7 +43,6 @@ import (
 	apiprometheus "github.com/grafana/grafana/pkg/services/ngalert/api/prometheus"
 	"github.com/grafana/grafana/pkg/services/ngalert/cluster"
 	"github.com/grafana/grafana/pkg/services/ngalert/eval"
-	"github.com/grafana/grafana/pkg/services/ngalert/folderlabel"
 	"github.com/grafana/grafana/pkg/services/ngalert/image"
 	"github.com/grafana/grafana/pkg/services/ngalert/metrics"
 	"github.com/grafana/grafana/pkg/services/ngalert/models"
@@ -201,7 +201,7 @@ type AlertNG struct {
 	tracer          tracing.Tracer
 	clientGenerator resource.ClientGenerator
 
-	folderLabel *folderlabel.Service
+	folderLabelReconciler *folderlabelsyncer.Service
 
 	evaluationCoordinator EvaluationCoordinator
 	schedCfg              schedule.SchedulerCfg
@@ -670,7 +670,7 @@ func (ng *AlertNG) init() error {
 
 	//nolint:staticcheck // not yet migrated to OpenFeature
 	if ng.FeatureToggles.IsEnabledGlobally(featuremgmt.FlagAlertingFolderHasRulesLabel) {
-		ng.folderLabel = folderlabel.NewService(ng.Cfg, ng.bus, ng.store, ng.clientGenerator)
+		ng.folderLabelReconciler = folderlabelsyncer.NewService(ng.Cfg, ng.bus, ng.store, ng.clientGenerator)
 	}
 
 	return ac.DeclareFixedRoles(ng.AccesscontrolService)
@@ -769,14 +769,14 @@ func (ng *AlertNG) Run(ctx context.Context) error {
 	})
 
 	//nolint:staticcheck // not yet migrated to OpenFeature
-	if ng.FeatureToggles.IsEnabledGlobally(featuremgmt.FlagAlertingFolderHasRulesLabel) && ng.folderLabel != nil {
+	if ng.FeatureToggles.IsEnabledGlobally(featuremgmt.FlagAlertingFolderHasRulesLabel) && ng.folderLabelReconciler != nil {
 		children.Go(func() error {
-			return ng.folderLabel.Run(subCtx)
+			return ng.folderLabelReconciler.Run(subCtx)
 		})
 		// Queues work onto the worker above, so it must start after it. Failures are logged and
 		// tolerated: a missed backfill leaves stale labels, it does not break rule evaluation.
 		children.Go(func() error {
-			if err := ng.folderLabel.Backfill(subCtx, ng.SQLStore, ng.store); err != nil {
+			if err := ng.folderLabelReconciler.Backfill(subCtx, ng.SQLStore, ng.store); err != nil {
 				ng.Log.Warn("Failed to backfill folder rules labels", "error", err)
 			}
 			return nil
