+++ pkg/services/ngalert/store/folderlabelsyncer/backfill_test.go
@@ -1,4 +1,4 @@
-package folderlabel
+package folderlabelsyncer
 
 import (
 	"context"
